Added hyprland configurations

+++ init.sh
@@ -1,15 +1,28 @@
 #!/usr/bin/bash
 
 base_dir=$(pwd)
-cp -as "$(pwd)/.config/" ~/
 
-sudo pacman -Syu - < i3packages
+read -p "This will replace existing configs in .config/. Are you shure?" -n 1 -r
+echo 
+if [[ $REPLY =~ ^[Yy]$ ]] then
+    # To lazy to create backups of already existing configs
+    cp -as "$(pwd)/.config/" ~/
+fi
+
+read -p "Install packages for i3 installation?" -n 1 -r
+echo 
+if [[ $REPLY =~ ^[Yy]$ ]] then
+    sudo pacman -Syu - < i3packages
+fi
 
 cd $HOME
+echo "Create links to tmux and vim config\n"
 ln -s "$base_dir/.tmux.conf" ~/
 ln -s "$base_dir/.tmux.conf.local" ~/
 ln -s "$base_dir/.vimrc" ~/
+mkdir -p ~/.vim/tmp
 
+echo "Create local bin folder"
 mkdir -p $HOME/.local/bin/
 echo "PATH=$HOME/.local/bin/:$PATH" >> ~/.bashrc
 cp $base_dir/systemMaintenance.fish $HOME/.local/bin/systemMaintenance
@@ -23,14 +36,14 @@ exec fish $LOGIN_OPTION
 fi' >> ~/.bashrc
 
 
-# Enable numlock after logged in
+echo "Enable numlock after logged in"
 echo "setleds -D +num" >> ~/.bash_profile
 
-# Download themes for rofi
+echo "Download themes for rofi"
 mkdir -p ~/.local/share/rofi/themes
 curl https://raw.githubusercontent.com/Rinfella/rofi-themes/refs/heads/master/arc_dark_transparent_colors.rasi -o ~/.local/share/rofi/themes/arc_dark_transparent_colors.rasi
 
-# Download plugin manager for vim
+echo "Download plugin manager for vim"
 curl -fLo ~/.vim/autoload/plug.vim --create-dirs \
     https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
